+++ Libplanet/Blockchain/BlockChain.cs
@@ -41,6 +41,11 @@ namespace Libplanet.Blockchain
         private readonly ILogger _logger;
 
         /// <summary>
+        /// Defines whether to allow rendering to be performed on this <see cref="BlockChain{T}"/>.
+        /// </summary>
+        private readonly bool _allowRender;
+
+        /// <summary>
         /// All <see cref="Block{T}"/>s in the <see cref="BlockChain{T}"/>
         /// storage, including orphan <see cref="Block{T}"/>s.
         /// Keys are <see cref="Block{T}.Hash"/>es and values are
@@ -69,15 +74,23 @@ namespace Libplanet.Blockchain
         /// it checks if the existing genesis block and this argument is the same.
         /// If the <paramref name="store"/> has no genesis block yet this argument will
         /// be used for that.</param>
+        /// <param name="allowRender">Defines whether to allow rendering to be performed on this
+        /// <see cref="BlockChain{T}"/>.</param>
         /// <exception cref="InvalidGenesisBlockException">Thrown when the <paramref name="store"/>
         /// has a genesis block and it does not match to what the network expects
         /// (i.e., <paramref name="genesisBlock"/>).</exception>
         public BlockChain(
             IBlockPolicy<T> policy,
             IStore store,
-            Block<T> genesisBlock
+            Block<T> genesisBlock,
+            bool allowRender = true
             )
-            : this(policy, store, store.GetCanonicalChainId() ?? Guid.NewGuid(), genesisBlock)
+            : this(
+                policy,
+                store,
+                store.GetCanonicalChainId() ?? Guid.NewGuid(),
+                genesisBlock,
+                allowRender)
         {
         }
 
@@ -85,14 +98,16 @@ namespace Libplanet.Blockchain
             IBlockPolicy<T> policy,
             IStore store,
             Guid id,
-            Block<T> genesisBlock
+            Block<T> genesisBlock,
+            bool allowRender = true
         )
             : this(
                 policy,
                 store,
                 id,
                 genesisBlock,
-                false
+                false,
+                allowRender
             )
         {
         }
@@ -102,7 +117,8 @@ namespace Libplanet.Blockchain
             IStore store,
             Guid id,
             Block<T> genesisBlock,
-            bool inFork
+            bool inFork,
+            bool allowRender = true
         )
         {
             Id = id;
@@ -111,6 +127,7 @@ namespace Libplanet.Blockchain
 
             _blocks = new BlockSet<T>(store);
             _transactions = new TransactionSet<T>(store);
+            _allowRender = allowRender;
             _rwlock = new ReaderWriterLockSlim(LockRecursionPolicy.SupportsRecursion);
             _txLock = new object();
 
@@ -841,7 +858,7 @@ namespace Libplanet.Blockchain
                 _rwlock.ExitUpgradeableReadLock();
             }
 
-            if (renderActions)
+            if (_allowRender && renderActions)
             {
                 RenderBlock(evaluations, block);
             }
@@ -1089,7 +1106,8 @@ namespace Libplanet.Blockchain
                     nameof(point));
             }
 
-            var forked = new BlockChain<T>(Policy, Store, Guid.NewGuid(), Genesis, true);
+            var forked = new BlockChain<T>(
+                Policy, Store, Guid.NewGuid(), Genesis, true, _allowRender);
             Guid forkedId = forked.Id;
             _logger.Debug(
                 "Trying to fork chain at {branchPoint}" +
@@ -1224,7 +1242,7 @@ namespace Libplanet.Blockchain
             _logger.Debug(
                 "Branchpoint is {branchPoint} (at {index})", topmostCommon, topmostCommon?.Index);
 
-            if (render)
+            if (_allowRender && render)
             {
                 _logger.Debug("Unrendering abandoned actions...");
 
@@ -1299,7 +1317,7 @@ namespace Libplanet.Blockchain
                 _rwlock.ExitWriteLock();
             }
 
-            if (render)
+            if (_allowRender && render)
             {
                 _logger.Debug("Rendering actions in new chain");